test(cuj): use app trust-badge contract harness in discovery tests
